Exercises 4.15-4.22

+++ Part4/bloglist/controllers/blogs.js
@@ -1,26 +1,65 @@
 const blogRoutes = require('express').Router()
 const Blog = require('../models/blogs')
 const mongoose = require('mongoose')
+const User = require('../models/users')
+const jwt = require('jsonwebtoken')
+
 mongoose.set('useFindAndModify', false)
 mongoose.set('useCreateIndex', true);
+
   
   blogRoutes.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
 
     response.json(blogs.map(blog => blog.toJSON()))
   })
+
+  const getTokenFrom = req => {
+    const authorization = req.get("authorization");
+    if (authorization && authorization.toLowerCase().startsWith("bearer ")) {
+      return authorization.substring(7);
+    }
+    return null;
+  };
+  
   
   blogRoutes.post('/', async (request, response) => {
     const body = request.body
+
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+
+    if(!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+  
+
+    if(!(body.url || body.title)) {
+      return response
+        .status(400)
+        .json({error: "Title is required"})
+        .end()
+    }
+
+    if(!body.likes) {
+      blog.likes = 0;
+    }
     
     const article = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes
+      likes: body.likes,
+      user: user._id
     })
   
     const savedBlog = await article.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.json(savedBlog.toJSON())
   })
 
@@ -50,9 +89,19 @@ mongoose.set('useCreateIndex', true);
   })
   
   blogRoutes.delete('/:id', async (request, response) => {
-   await Blog.findByIdAndRemove(request.params.id)
+    
 
-   response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+
+    const blog = await Blog.findById(request.params.id)
+    const blogId = blog.user.toString()
+    const userId = decodedToken.id
+
+    if (blogId === userId) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
   })
 
   module.exports = blogRoutes
